kill and remove Grafana and Prometheus docker instances without error code in case they do not exist

+++ kill-grafana.sh
@@ -26,5 +26,10 @@ else
     GRAFANA_NAME=agraf-$GRAFANA_PORT
 fi
 
-sudo docker kill $GRAFANA_NAME
-sudo docker rm -v $GRAFANA_NAME
+if [ "$(sudo docker ps -q -f name=$GRAFANA_NAME)" ]; then
+    sudo docker kill $GRAFANA_NAME
+fi
+
+if [[ "$(sudo docker images -q $GRAFANA_NAME 2> /dev/null)" != "" ]]; then
+    sudo docker rm -v $GRAFANA_NAME
+fi
